MTV-2156: Remove eslint-plugin
