+++ .prettierrc.js
@@ -1,0 +1,9 @@
+/* eslint-env node */
+module.exports = {
+    printWidth: 100,
+    tabWidth: 2,
+    semi: true,
+    singleQuote: true,
+    trailingComma: 'all',
+    bracketSpacing: true,
+};
