+++ .stylelintrc.yml
@@ -1,0 +1,20 @@
+extends:
+  - stylelint-config-standard
+  - stylelint-config-prettier
+rules:
+  indentation:
+    - 2
+  color-named: never
+  color-no-hex: true
+  custom-property-pattern: null
+  function-disallowed-list:
+    - rgb
+  selector-class-pattern: null
+  selector-disallowed-list:
+    - '*'
+    - /\.(pf|co)-/
+  selector-max-type:
+    - 0
+    - ignore:
+        - compounded
+        - descendant
