+++ cspell.wordlist.txt
