--- packages/common/.eslintrc.cjs
@@ -1,6 +1,0 @@
-/* eslint-env node */
-
-module.exports = {
-  root: true,
-  extends: ['plugin:@kubev2v/eslint-plugin/typescript'],
-};
