--- packages/common/.prettierrc.cjs
@@ -1,5 +1,0 @@
-/* eslint-env node */
-
-module.exports = {
-  ...require('@kubev2v/eslint-plugin/prettierrc.js'),
-};
