--- packages/common/.stylelintrc.yml
@@ -1,1 +1,0 @@
-extends: "../eslint-plugin/stylelintrc.yml"
