--- packages/eslint-plugin/README.md
@@ -1,3 +1,0 @@
-# @kubev2v/eslint-plugin
-
-Collection of linting configurations shared by `@kubev2v` packages.
