--- packages/eslint-plugin/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@kubev2v/eslint-plugin",
-  "version": "0.0.1",
-  "description": "standard eslint configuration for forklift",
-  "license": "Apache-2.0",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/kubev2v/forklift-console-plugin.git",
-    "directory": "packages/eslint-plugin-forklift"
-  },
-  "main": "index.js",
-  "scripts": {
-    "lint": "eslint .",
-    "lint:fix": "eslint --fix .",
-    "test": "echo no tests yet",
-    "test:coverage": "echo no tests yet",
-    "test:updateSnapshot": "echo no tests yet"
-  },
-  "devDependencies": {
-    "eslint": "^8.0"
-  }
-}
