--- packages/eslint-plugin/prettierrc.js
@@ -1,8 +1,0 @@
-/* eslint-env node */
-
-module.exports = {
-  arrowParens: 'always',
-  singleQuote: true,
-  trailingComma: 'all',
-  printWidth: 100,
-};
