--- packages/forklift-console-plugin/.stylelintrc.yml
@@ -1,1 +1,0 @@
-extends: "../eslint-plugin/stylelintrc.yml"
