--- testing/.prettierrc.cjs
@@ -1,5 +1,0 @@
-/* eslint-env node */
-
-module.exports = {
-    ...require("../packages/config/prettierrc"),
-};
